Adds testnet check for account
Improves user experience by adding a check to ensure the user is on the correct network (testnet) before interacting with the dApp.
This prevents errors and provides a warning message if the user is on the wrong network.
Also disables the 'Reclaim' and 'Claim prize' buttons when the user is on the wrong network.